Configure `easy_install` to use an `index_url` if pip is
One of the synapse dependencies, Twisted, uses `easy_install` to install
its own dependencies. Unfortunately `easy_install` does not honour the
`index_url` used by pip (if one is used). As a slight hack we try to
extract the `index_url` used by pip and put it in a configuration file
that `easy_install` will read to set the `index_url`.