Port homeserver.yaml.jinja to v0.99.0
Update the homeserver config template to be reflect a config file that
would be generated by Synapse v0.99.0. This should help to avoid any
potential breakage as Synapse is upgraded.

Support for `tls_dh_params_path` was removed in:

    https://github.com/matrix-org/synapse/commit/23b0813599

so the `tls_dh_file` config value is no longer needed in `defaults.yaml`